requirements.txt
pandas
duckdb
pyarrow
requests
pyyaml
selfies
CoolProp
rdkit
pytest
beautifulsoup4
html5lib
lxml
